catfile: Fix race with done-notifications
The notification that a process is done and closed gets sent out right
at the beginning of `returnWhenDone()` and is thus sent out _before_ the
process really is done already. This thus doesn't quite signal that the
process is done, but just that is about to be processed.

Defer the signal to the end such that it's sent only when the process
has really been processed.