Intercept RepositoryExists only with PerRepository as election strategy
RepositoryExists was changed to be intercepted in Praefect regardleess of
the election strategy in abdd0e85. Previously it was only intercepted with
'per_repository' election strategy configured. This should be fine in real
use as the other election strategies are not configurable outside of tests.
Some tests in GitLab's repository still depend on the old election strategies
and do not set up a database. To retain compatibility with the said tests and
to reintroduce the changes reverted in fdda523b, this commit brings back the
condition to only intercept the RPC if 'per_repository' is the configured
election strategy.