catfile: Allow reading objects with unknown type
The `objectReader` structure currently requires the caller to know about
the object's type before it is actually being read: if the type does not
match, then the object's data is discarded and an error is returned.
This forces callers to first learn about the object type, e.g. via `git
cat-file --batch-info`, which is additional overhead that is often not
necessary at all.

Refactor the interface to not require a type anymore. Like this, callers
can use the interface without knowing about the object type beforehand.
Instead, callers may still discard the object in case they realize it is
not the type they expected.