updateref: Fix test flakiness when writing to updater fails
In one of our tests for the updateref package, we verify that we get an
error when trying to commit a change on a transaction that has already
been prepared. The test is flaky though given that it sometimes does not
return the expected error message from git-update-ref(1). This race is
caused by git-update-ref(1) exiting before we get to the point where we
can write another status change command to its stdin, and as a result we
get a failure writing to the command's pipe instead of the expected
error.

Fix this bug by explicitly cancelling the command and including its
stderr when writing to it fails. This makes sure that stderr was flushed
and we've got it available, and it would also include the error message
we are searching for in the test.