Backfill replica_path in 'repositories' records
Praefect now generates a repository id that can be used to join the
database records of a repository across the tables. As such, the
virtual_storage and relative_path columns will not be needed in other
tables than the 'repositories' table as the joining will happen via the
repository_id column. As Praefect will begin generating unique relative
paths for the replicas to avoid stale disk state of deleted repositories
affecting recreation of said repositories, the replica_path column was
added to the 'repositories' table to store the actual disk path of the
replicas. Right now, every newly created 'repositories' record has the
replica_path set to the relative_path. To begin using the new column,
we also need the historical records to have the column correctly filled.
This commit adds a migration that fills the replica_path of existing
records. As each repository is currently stored in the path sent by the
client, we'll just fill the column using that.

Changelog: other