updateref: Always assert state transitions if Git supports flushing
With 4abf27024 (updateref: Fix indeterminate state when locking refs,
2021-09-13), we have introduced proper verification of state transitions
in git-update-ref(1) such that we notice early on when locking of refs
fails. As a result, we won't try to vote on transactions where we could
have already known that git-update-ref(1) failed to lock refs in the
first place and would thus abort early on.

Remove the feature flag guarding this code. Given that this requires
changes in Git, we still need to verify that the current Git version
supports flushing of status updates.