catfile: Plug Goroutine leak in the cache
When constructing processes as part of the catfile cache, then we must
create a separate context in case the process is cacheable such that it
does not get killed when the original request context gets cancelled.
Starting with commit 36c166a17 (catfile: Explicitly close processes,
2021-09-09), we have converted the code to always explicitly close
processes instead of doing so via the context. As a result though, this
separate detached context will never get cancelled at all, causing us to
leak Goroutines which listen on context cancellation to perform various
cleanup operations.

While we could simply revert this commit, the original implementation
has been quite indirect given that all process cancellation happened via
this context. The context has thus been passed to the `BatchProcess`,
which knew to cancel it as required. This is the burden to children of
the cache, which is an inversion of responsibilites: it should be the
cache which is managing lifetimes, not the processes.

Fix the Goroutine leak by instead stowing away the separate context's
cancellation function as part of the cache entries. Like this, the cache
can retrieve and cancel them whenever it evicts entries from the cache,
and it can cancel them when a process cannot be returned to the cache in
case it is empty. In the other case of uncacheable processes, we do not
replace the surrounding context's done function at all and can thus
continue to rely on the caller to eventually cancel the context.

Fixes: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5566
Changelog: fixed