supervisor: Stop monitor if shutdown was signalled
The supervisor's health monitor will check health every 15 seconds or
alternatively exit if its shutdown channel was closed. While we do
verify this channel correctly when trying to send the event, we do not
check for shutdown when waiting 15 seconds for the next event.

Fix this by returning early if we got the signal to shut down.