Support injecting router in to runPraefectServer test helper
runPraefectServer is a helper to launch a Praefect server in tests.
By default, it sets up a NodeManagerRouter for the Praefect server
that is launched. This commit makes it possible to provide a custom
router instead to make it possible to provide a PerRepositoryRouter
instead in selected tests. Longer term, we should migrate all of the
tests to use PerRepositoryRouter instead as NodeManagerRouter is no
longer usable.