cache: Add ability to stop disk walkers
The disk cache spawns a set of walkers which regularly walk the cache to
check for cache members which are older than a given threshold. These
walkers currently cannot be stopped at all, leading to a Goroutine
leakage in our tests.

Implement a new `StopWalkers()` function which stops them and use it in
our tests.