glsql: Close database if open fails on ping
When opening the database, we first create the connection and then ping
it to assert that we can correctly establish the connection. If the ping
fails though, then we leak the connection because we don't clean up the
already-opened database struct.

Fix this issue by closing the database if the ping failed.