praefect: Stop gRPC servers on exit
Next to a bunch of other things, the Praefect server's `run()` function
also has the task to create the gRPC servers. These servers are never
stopped though, which as a result means that we leak listeners set up
for this server in our tests.

Fix this by calling `Stop()` on these servers.