glsql: Refactor database setup to not modify environment
The database setup in the glsql package uses `gdk env` to get the
database configuration from the GDK such that users do not have to
manually expose these envvars before running tests. Currently though, we
use the output of `gdk env` to modify the tests' own environment, which
is bad practice as it can have unintended sideeffects.

Refactor the code to instead just pull out required environment
variables and pass it to the config. Furthermore, execute `gdk env` once
only such that we do not have to pay the cost of spawning it repeatedly
for every test that needs a database configuration.