updateref: Fix test flakiness depending on when git-update-ref(1) exits
Similar to 88b946a0d (updateref: Fix test flakiness when writing to
updater fails, 2021-09-21), we have another flaky test depending on when
git-update-ref(1) fails: when writing to it, or when trying trying to
commit.

Fix the flakiness by checking for a smaller subset of the error message:
what we care about is that Git fails while locking, and this error
message is present in both error paths.