sidechannel: remove waiter.Wait()
Before this change, the client-side sidechannel.Waiter object had a
Close() and a Wait() method. These methods were the same except
Close() would _sometimes_ block, and Wait() would _always_ block
waiting for the sidechannel callback to run. This does not make sense
when implementing a sidechannel proxy (like we need for Praefect)
because how does the proxy know if it needs to call Wait() or Close()?

This change adds an extra message to the sidechannel protocol: the
client sends two bytes "ok" to the server after accepting the
sidechannel connection. This creates a new synchronization point. Now,
RPC's that want the behavior of Wait() can implement it by first
blocking in the client waiting for a gRPC message from the server, and
then calling Waiter.Close(). This makes proxying easier because a
proxy can now just call Close() after the gRPC call is done.

This is a breaking change to the protocol but that is still OK because
we have no RPC's yet that use the protocol.

Changelog: other