catfile: Disentangle metrics from managing the cache
We will split up the cache into two caches, one for object readers and
one for object info readers. For this, we will want to reuse the "stack"
that keeps track of cached processes. But due to it currently being
entangled with Prometheus metrics, this refactoring is harder to do than
one would want it to be.

Pull up metrics into the high-level functions such that we can easily
split out the low-level stack at a later point. This also has the
benefit that we're not handling metrics in the critical section anymore.