catfile: Convert reporting of cache members to use a gauge vector
We will split up the cache into two caches, one for object readers and
one for object info readers. As a result, the current gauge we use for
tracking the number of cached batch processes will not be as informative
anymore given that it can only report the total count of processes, but
cannot distinguish the types.

Convert the gauge to a vector with a "type" label such that we can
distinguish cached processes. While at it, this also pulls out handling
of the metric from generic parts which manage the stack of cached
processe and instead moves it into a separate function. This has the
benefit that we can easily split out the generic logic for the second
cache, and it removes handling of metrics from the critical section.