catfile: Make `close()` a private function
We do not allow callers outside of the catfile package to call `Close()`
on returned processes given that we want to be the only ones which
manage the lifetime of potentially cached processes.

Make the function private to better highlight this. This allows us to
embed it into the reader interfaces without exposing close to outside
callers such that we can easily access the `close()` function on
interfaces internally.