Expose repository_id from valid_primaries view
valid_primaries view contains the replicas which are eligible to act
as the primary for each repository. This commit updates the view to expose the
repository_id so the queries can be updated in Praefect to perform their joins
using it. The view itself is not yet updated as some tests will fail backwards
compatibility checks as they do not setup the repository IDs correctly. This is
fine as exposing the id allows us to update the queries which could be still
ran by outdated Praefects when the upgrade to 14.5 is happening. The view itself
can be transactionally upgraded via a migration so it won't cause problems
during the upgrade. We'll have to update the view to perform the joins via
the repository ID in 14.5.