Wait for the first health check prior to serving in Praefect
Praefect currently is not waiting for a health check to complete
before it takes over the sockets and starts routing RPCs. This can
be a problem as the RPCs will fail due to not having any healthy nodes
available if they are served prior to performing the first check. This
commit fixes the issues by waiting for the first health check to complete
prior to serving RPCs.

Changelog: fixed