Skip NamespaceService tests when testing with Praefect
Praefect's PerRepositoryRouter doesn't support storage scoped mutators.
This fails NamespaceService tests exercising RPCs which are classified
as such. In practice, this is not a problem as the RPCs are not being
called anymore with hashed storage in use. The RPCs in NamespaceService
should be removed but likely we can't do that quite yet as some tests
in GitLab use them. This commit worksaround the problem by skipping the
tests that fail if Praefect is in use. This allows us to move forward
with enabling 'per_repository' election strategy in our test-with-praefect
target to better test the actual production stack. Later, we should drop
the NamespaceService RPCs from Gitaly once the tests in Rails are dropped
as well.