Don't use forceCreateRepositories hack for repo creations
Repository creations fail with Praefect in tests due to the testing
hack that creates repository records. As the hack creates a record
for the repository before the actual call is executed, the actual
call will fail with 'repository exists'. This commit worksaround the
problem by not using the hack for repository creation calls. Ideally
we would not have such a hack in the first place and would use the
API to create the repositories in the tests.