backup: Use CreateRespository and FetchBundle to restore bundles
To restore from an incremental backup we will need to apply multiple
bundles to the same repository. This is not possible with
CreateRepositoryFromBundle. So instead we first create an empty
repository and then fetch the contents of each bundle into it.

To maintain compatibility with existing backups we introduce the idea of
"skippable" bundles. It wouldn't make sense to skip a whole restore when
an incremental bundle doesn't exist. So here we're making sure we only
skip non-existing full backups.