Accept patterns for BundleTestRepo test helper
Previously this helper would create bundles for all refs in a
repository but to validate incremental backups we will need to be able
to generate bundles that only have a selection of refs/objects.

To maintain compatibility with existing call sites the default behaviour
(no patterns) will continue to bundle all refs.