replication: Configure number of replication workers per virtual storage
The new configuration parameter to set number of the workers used
to process replication jobs. It is a safe incremental move on to the
rolling out parallel storage processing by the replication manager.
The default value 1 behaves the same as no changes done to replication
manager (the old behaviour).

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2915