list-untracked-repositories: Change callback of ExecOnRepositories to return an error
The ExecOnRepositories method accepts a callback function that
is called once the batch is ready for the processing. The signature
of the callback function changed to return an error. If the error
occurs the ExecOnRepositories stops iteration over the repositories
and returns an error back to the caller. That change is a preparation
for the new 'list-untracked-repositories' sub-command.