sql-migrate: Update storage_repositories table
The storage_repositories table has triggers registered
on table insert and table update operations. The trigger
create a JSON that contains the change done to each row
and send notification with JSON as a payload. Because of
the PostgreSQL limitations the payload can't be too big.
That is why we do migration for the storage_repositories
table in batches. Each batch produces fewer data than
allowed by the limit and notifications could be sent without
any issues.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3806

Changelog: fix