backup: Adds previous ref path to backup steps
When creating an incremental backup we need access to the previous
increments references. Adding this to Step allows each step to be as
isolated as possible such that we can create an incremental based on a
single step only, similar to how we can create a full backup based on a
single step.