testhelper: Move process leak checks into separate package
The function `MustHaveNoChildProcess()` is not a general test helper
that should be used by tests as they see fit, but is instead intended to
be run at the end of tests only. Move it into a separate package and
make it internal to the package to make it less discoverable by callers.