testhelper: Enable Goroutine leak checks by default
In a recent incident, a refactoring of the catfile cache has caused an
incident due to leaking Goroutines. While we did already check for
leaking Goroutines in a subset of our test suite before via the goleak
check, this was opt-in the same way as `mustHaveNoChildProcess()` was.
It also was used a lot less often given that we did have hundreds of
actual Goroutine leaks in our test suite.

Introduce leak checking by default now that almost all of the Goroutine
leaks have been fixed. There is only two packages remaining which still
leak Goroutines: the catfile package, which will get fixed via an MR
which is currently in flight. And the backup package, which is leaking
Goroutines in a package external to us. To the best of my knowledge,
there is no way to plug those.

In any case, having on-by-default leak checking should put us in a much
better position to detect such regressions in the future.