testhelper: Return errors when configuring Ruby sidecar fails
When configuration of the Ruby sidecar fails, then we call
`log.Fatalf()` and thus directly execute the complete test suite. Given
that we call this function on-demand in tests nowadays though, we should
instead return the error and validate at the test site that it did set
up the server correctly such that other tests continue to run as
expected.

Fix this issue by returning the error instead of aborting.