catfile: Provide function to safely count cache members
The `len()` function provided by the `ProcessCache` structure counts
entries without taking a lock, which isn't all that helpful: we can just
call `len()` directly. The more interesting case though is to count
cache members where we're required to take the lock due to possible
concurrent updates, and we're about to add a few callsites which need
this.

Remove the old `len()` function and replace it with `entryCount()`,
which counts members under the mutex. This is a preparatory commit for
the subsequent commit which changes the cache members Promeutheus metric
into a vector.