Add GetReplicaPath to PostgresRepositoryStore
This commit adds GetReplicaPath to PostgresRepositoryStore. It retrieves
the storage path where the replicas of the repository are stored from the
database. This needs to be done in 14.5 to retain backwards compatibility
when Praefect begins to generate unique disk paths in 14.6. The method
will be used when applying replication jobs and when routing requests to
direct the RPCs to the correct locations.