Use repository id in RepositoryStoreCollector
RepositoryStoreCollector needs to join the database records via the
repository ID in order to be ready when Praefect begins generating
unique relative paths for repositories and drops the virtual storage
and relative path columns from storage_repositories and repository_
assignments. This commit updates the query to join the records using
the repository id.