nodes: Fix leaking connections and Goroutines due to missing cleanup
The node manager creates connections to all of its known nodes and
starts monitoring routines which check the respective nodes' health
status. We never clean up either of them, which thus leads to lots of
Goroutine leakages in our tests.

Fix this by providing a new `Stop()` function for the manager which both
stops the electors' monitoring Goroutines and closes the node
connections and call this function as required.