testhelper: Simplify infra to build commands once
In order to not have to repeatedly rebuild the same executables during
test setups, the testhelper package provides a bunch of functions which
build executables via `sync.Once` and then copy them into place. One
shortcoming though is that we have a separate `sync.Once` for each of
the binaries though, which is easy to get wrong.

Refactor the code to use a `sync.Map` containing `sync.Once`s keyed by
source path. Like this, we can create them ad-hoc in a raceless way and
don't need to manage them globally.