supervisor: Close writers before signalling process is done
When a supervised process is going down, then we need to close its
associated stdout and stderr loggers. This is done after we signal to
the caller that the process is done though by closing the wait channel,
which in turn means that any Goroutines spawned by the writers may still
be active at the time we're exiting.

Fix this by instead deferring the close. While this fixes the issue, it
also avoids any deadlocks due to an unclosed wait channel in case any of
the invoked function calls panics.