testhelper: Refactor building of binaries for improved readability
The testhelper function to build binaries is a bit hard to read. One of
the sources is the variable names which point to the shared binary
paths, which are rather generically named and thus don't tell what these
really are about. And the other confusing part is the deferred function
call, which isn't really necessary at all.

Improve readability by renaming the variables and inlining the deferred
call.