testserver: Clean up Praefect proxy if it fails to become healthy
The Praefect proxy is started via a separate executable which we'll
automatically inject when a certain environment variable is set. If the
Praefect proxy fails to come up and become healthy though, we leak this
executable because we don't kill it if the test fails. As a result,
there is a misleading panic at the end of the test run hinting at a
leaked process, while the real issue is that Praefect failed to become
healthy at all.

Fix this by queueing shutdown of the process before checking for
Praefect to become healthy.