testserver: Drop retry when waiting for server to become healthy
When executing `waitHealthy()`, then we retry connecting to the health
client, where each retry is limited to 1 second. This shouldn't be
necessary though:

    - `grpc.WaitForReady()` ensures that we retry connecting to the
      service until we're either successful or until the context times
      out. It's thus useless to re-dial.

    - When instantiating the health server, it will always be
      initialized in "serving" state for the default service.
      Furthermore, we only ever check for the default service and never
      modify the serving status of the server. As such, if we ever
      receive an answer from the server, then it must be that it is
      serving. It's thus useless to retry the RPC call.

In combination, it means that we do not need to loop at all in
`waitHealthy()`.

Simplify the function to not loop, but instead use a 3 second timeout
both for dialing the server and executing the health check.