testserver: Always assert that Gitaly can serve requests
When starting Gitaly with Praefect, then we always wait until Gitaly
becomes healthy before we return from the function such that Praefect
can correctly route information to it. It does make sense to wait for
Gitaly to be healthy in the general case though such that we don't ever
run with Gitaly not yet being able to server requests.

Improve the situation by always asserting that Gitaly is healthy when
`RunGitalyServer()` or `StartGitalyServer()` returns. This is likely to
fix some flaky tests where we have seen Gitaly to not be able to serve
requests.