testhelper: Reject trying to build same binary multiple times
We allow for the case where a test tries to build a binary with the same
target path multiple times. This is bound to be problematic though: if
we ever were to blindly parallelize these tests, then we'd potentially
copy over the target path while it's still being executed, which would
likely lead to ETXTBSY errors. It's furthermore an antipattern to reuse
the same configuration for multiple test runs due to potentially shared
state.

Fix all tests which do this. Next to a few tests which accidentally
build gitaly-hooks multiple times, the most common case is the
"praefect" binary which gets built automatically whenever we run a
new Gitaly server. Add a safeguard which fails the test as soon as it
tries to build the same binary twice.