git: Remove `Config` interface
The `Config` interface is not used anymore except for in our tests.
Furthermore, no new callsites should be added given that we nowadays
have `localrepo.SetConfig()` and `localrepo.UnsetMatchingConfig()`, both
of which know to correctly lock, vote on and commit the changed config
in a racefree way.

Remove the interface and the localrepo implementation of it.