migrations: Fix data race due to accessing global state
We have recently hit a data race in our migration tests due to two
different Goroutines accessing the global `IgnoreUnknown` flag in the
sql-migrations package. As it turns out, sql-migrations provides two
different APIs: a "simple" API which uses a global `MigrationSet` struct
to keep data in. And an explicit API where the caller needs to create
the `MigrationSet` by himself. Given that we use the former, it's clear
that we easily run into races as soon as two concurrent tests want to do
migrations.

Fix the data race by converting all callsites to use an explicit
`MigrationSet` structure without any global state.