ci: Fix backwards-compatibility job with changing setup
The database backwards-compatibility job is testing whether an old
version of Praefect continues to work correctly in case it's running
with a database which has been migrated to a newer schema. To do so, we
check out the previous minor release, but restore migrations from the
current version. While this works alright in case there are only changes
to the migrations themselves (e.g. only migrations have been added), it
falls apart when the setup of migrations itself changes.

An upcoming change will remove setup of the migration package in
"migrations.go" and move it into local state. Naturally, if we run an
older version which does not yet set up that local state in the other
packages with a "migrations.go" file which doesn't set up the table name
anymore, the only result can be failure.

Fix the issue by not modifying "migrations.go" anymore.